+++ src/Exception/UnsupportedMessageStructureException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Beter\Yii2BeterLogging\Exception;
+
+class UnsupportedMessageStructureException extends \Exception
+{
+}