+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 [[#1](https://github.com/BETER-CO/yii2-beter-logging/issues/1)].
 - Redundant new line before `log.trace` output in the `console` formatter.
 - Improvement in readability of th `console` formatter output.
+- `ProxyLogTarget` supports delayed errors with a context.
 
 ### 1.0.0